+++ src/Console/Commands/MakeModuleModel.php
@@ -40,14 +40,18 @@ class MakeModuleModel extends Command
     {
         $module_name = $this->ask('Enter your module name');
         if(File::exists('modules/' . $module_name)) {
-            $model_name = $this->ask('Enter your model name');
-            $this->call('make:model', ['name' => $model_name]);
             $model_route = "modules/" . $module_name . "/src/Models/";
-            if(!File::exists($model_route)) {
-                File::makeDirectory($model_route,0777,true);
+            $model_name = $this->ask('Enter your model name');
+            if (!File::exists($model_route . $model_name . '.php')){
+                $this->call('make:model', ['name' => $model_name]);
+                if(!File::exists($model_route)) {
+                    File::makeDirectory($model_route,0777,true);
+                }
+                $command = "mv app/Models/$model_name" . ".php $model_route";
+                exec($command);
+            }else{
+                $this->error('Model already exists.');
             }
-            $command = "mv app/Models/$model_name" . ".php $model_route";
-            exec($command);
         }else{
             $this->error('Module does not exist.create the module using module:make command.');
         }
